requirements.txt
soundfile
librosa
pytorch
pytorch_lightning
asteroid
asteroid_filterbanks